+++ libs/tx-builder/src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './search';
 export * from './abi';
 export * from './cache';
 export * from './decoding';
+export * from './ethersFallback';
