+++ libs/utils/src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './shamanPermission';
 export * from './time';
 export * from './validation';
 export * from './error';
+export * from './ethersFallback';
